+++ src/App/index.tsx
@@ -1,6 +1,11 @@
 import React from 'react';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
-import {HeaderController, Login, HomeController, Detail} from '@/components';
+import {
+  HeaderController,
+  Login,
+  HomeController,
+  DetailController,
+} from '@/components';
 
 const App = () => {
   return (
@@ -10,7 +15,7 @@ const App = () => {
         <Routes>
           <Route path="/" element={<Login />} />
           <Route path="/home" element={<HomeController />} />
-          <Route path="/detail/:id" element={<Detail />} />
+          <Route path="/detail/:id" element={<DetailController />} />
         </Routes>
       </Router>
     </div>
