+++ src/components/Detail/DetailElements/Controls/Controls.tsx
@@ -1,0 +1,130 @@
+import React from 'react';
+import styled from 'styled-components';
+import {IMAGES} from '@/assets';
+
+export const Controls = () => {
+  return (
+    <Container>
+      <PlayerButton>
+        <img src={IMAGES.PLAY_ICON_BLACK_URL} alt="" />
+        <span>Play</span>
+      </PlayerButton>
+      <TrailerButton>
+        <img src={IMAGES.PLAY_ICON_WHITE_URL} alt="" />
+        <span>Trailer</span>
+      </TrailerButton>
+      <AddListButton>
+        <span />
+        <span />
+      </AddListButton>
+      <GroupWatch>
+        <div>
+          <img src={IMAGES.GROUP_ICON_URL} alt="" />
+        </div>
+      </GroupWatch>
+    </Container>
+  );
+};
+
+const Container = styled.div`
+  align-items: center;
+  display: flex;
+  flex-flow: row nowrap;
+  margin: 24px 0px;
+  min-height: 56px;
+`;
+
+const PlayerButton = styled.button`
+  font-size: 15px;
+  margin: 0px 22px 0px 0px;
+  padding: 0px 24px;
+  height: 56px;
+  border-radius: 4px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  letter-spacing: 1.8px;
+  text-align: center;
+  text-transform: uppercase;
+  background: rgb(249, 249, 249);
+  border: none;
+  color: rgb(0, 0, 0);
+
+  img {
+    width: 32px;
+  }
+
+  &:hover {
+    background: rgb(198, 198, 198);
+  }
+
+  @media (max-width: 768px) {
+    height: 45px;
+    padding: 0px 12px;
+    font-size: 12px;
+    margin: 0px 10px 0px 0px;
+
+    img {
+      width: 25px;
+    }
+  } ;
+`;
+
+const TrailerButton = styled(PlayerButton)`
+  background: rgba(0, 0, 0, 0.3);
+  border: 1px solid rgb(249, 249, 249);
+  color: rgb(249, 249, 249);
+`;
+
+const AddListButton = styled.div`
+  margin-right: 16px;
+  height: 44px;
+  width: 44px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.6);
+  border-radius: 50%;
+  border: 2px solid white;
+  cursor: pointer;
+
+  span {
+    background-color: rgb(249, 249, 249);
+    display: inline-block;
+
+    &:first-child {
+      height: 2px;
+      transform: translate(1px, 0px) rotate(0deg);
+      width: 16px;
+    }
+
+    &:nth-child(2) {
+      height: 16px;
+      transform: translateX(-8px) rotate(0deg);
+      width: 2px;
+    }
+  }
+`;
+
+const GroupWatch = styled.div`
+  height: 44px;
+  width: 44px;
+  border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  background: white;
+
+  div {
+    height: 40px;
+    width: 40px;
+    background: rgb(0, 0, 0);
+    border-radius: 50%;
+
+    img {
+      width: 100%;
+    }
+  }
+`;
