+++ src/components/Detail/DetailElements/index.ts
@@ -1,0 +1,1 @@
+export * from './Controls/Controls';
