+++ src/components/Detail/DetailLayout.tsx
@@ -1,0 +1,96 @@
+import React from 'react';
+import styled from 'styled-components';
+import {MovieType, Nullable} from '@/types';
+import {Controls} from './DetailElements';
+
+type DetailLayoutProps = {
+  data: Nullable<MovieType>;
+};
+
+export const DetailLayout = ({data}: DetailLayoutProps) => {
+  return (
+    <Container>
+      <Background>
+        <img src={data?.backgroundImg} alt={data?.title} />
+      </Background>
+
+      <ImageTitle>
+        <img src={data?.titleImg} alt={data?.title} />
+      </ImageTitle>
+
+      <ContentMeta>
+        <Controls />
+        <SubTitle>{data?.subTitle}</SubTitle>
+        <Description>{data?.description}</Description>
+      </ContentMeta>
+    </Container>
+  );
+};
+
+const Container = styled.div`
+  position: relative;
+  min-height: calc(100vh-250px);
+  overflow-x: hidden;
+  display: block;
+  top: 72px;
+  padding: 0 calc(3.5vw + 5px);
+`;
+
+const Background = styled.div`
+  inset: 0px 0px 0px;
+  opacity: 0.8;
+  position: fixed;
+  z-index: -1;
+
+  img {
+    width: 100vw;
+    height: 100vh;
+
+    @media (max-width: 768px) {
+      width: initial;
+    }
+  }
+`;
+
+const ImageTitle = styled.div`
+  align-items: flex-end;
+  display: flex;
+  -webkit-box-pack: start;
+  justify-content: flex-start;
+  margin: 0px auto;
+  height: 30vw;
+  min-height: 170px;
+  padding-bottom: 24px;
+  width: 100%;
+
+  img {
+    max-width: 600px;
+    min-width: 200px;
+    width: 35vw;
+  }
+`;
+
+const ContentMeta = styled.div`
+  max-width: 854px;
+`;
+
+const SubTitle = styled.div`
+  color: rgb(249, 249, 249);
+  font-size: 15px;
+  min-height: 20px;
+
+  @media (max-width: 768px) {
+    font-size: 12px;
+  }
+`;
+
+const Description = styled.div`
+  line-height: 1.4;
+  font-size: 20px;
+  padding: 16px 0px;
+  color: rgb(249, 249, 249);
+
+  @media (max-width: 768px) {
+    font-size: 14px;
+  }
+`;
