+++ src/components/Header/HeaderElements/NavMenu/navMenuInfo.map.ts
@@ -8,31 +8,31 @@ export const navMenuInfo = [
     title: 'HOME',
   },
   {
-    path: '/',
+    path: '/home',
     imgUrl: IMAGES.SEARCH_ICON_URL,
     alt: 'SEARCH',
     title: 'SEARCH',
   },
   {
-    path: '/',
+    path: '/home',
     imgUrl: IMAGES.WATCHLIST_ICON_URL,
     alt: 'WATCHLIST',
     title: 'WATCHLIST',
   },
   {
-    path: '/',
+    path: '/home',
     imgUrl: IMAGES.ORIGINAL_ICON_URL,
     alt: 'ORIGINALS',
     title: 'ORIGINALS',
   },
   {
-    path: '/',
+    path: '/home',
     imgUrl: IMAGES.MOVIE_ICON_URL,
     alt: 'MOVIES',
     title: 'MOVIES',
   },
   {
-    path: '/',
+    path: '/home',
     imgUrl: IMAGES.SERIES_ICON_URL,
     alt: 'SERIES',
     title: 'SERIES',
