+++ src/components/Header/HeaderElements/index.ts
@@ -1,0 +1,1 @@
+export * from './NavMenu/NavMenu';
