+++ src/components/Home/HomeElements/ImgSlider/ImgSlider.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 import Slider from 'react-slick';
+import styled from 'styled-components';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import {imgSliderInfo} from './imgSliderInfo.map';
