+++ src/components/Home/HomeElements/ImgSlider/imgSliderInfo.map.ts
