+++ src/components/Home/HomeElements/Trending/Trending.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import {useSelector} from 'react-redux';
-import {selectOriginal} from '@/store/movies/movies.selectors';
+import {selectTrending} from '@/store/movies';
 import {MoviesLayout} from '@/components';
 
-export const Originals = () => {
+export const Trending = () => {
   //Variables
   const title = `Originals`;
 
   //Selectors
-  const movies = useSelector(selectOriginal);
+  const movies = useSelector(selectTrending);
 
   return <MoviesLayout title={title} data={movies} />;
 };
