+++ src/components/Home/HomeElements/index.ts
@@ -1,0 +1,7 @@
+export * from './ImgSlider/ImgSlider';
+export * from './Viewers/Viewers';
+export * from './Recommends/Recommends';
+export * from './NewDisney/NewDisney';
+export * from './Originals/Originals';
+export * from './Trending/Trending';
+export * from './MoviesLayout/MoviesLayout';
