+++ src/components/Login/Login.tsx
@@ -13,7 +13,6 @@ export const Login = () => {
             Get Premier Access to Raya and the Last Dragon for an additional fee
             with a Disney+ subscription. As of the bloody date of 24/02/2022,
             the price of Disney+ and The Disney Bundle will increase by $1.
-            Should this information come from database?
           </Description>
           <CTALogoTwo src={IMAGES.CTA_LOGO_TWO_URL} alt="" />
         </CTA>
