+++ src/components/index.ts
@@ -1,13 +1,5 @@
 export * from './Login/Login';
 export * from './Header/HeaderController';
-export * from './Header/Header';
-export * from './NavMenu/NavMenu';
+export * from './Detail/DetailController';
 export * from './Home/HomeController';
-export * from './ImgSlider/ImgSlider';
-export * from './Viewers/Viewers';
-export * from './MoviesLayout/MoviesLayout';
-export * from './Recommends/Recommends';
-export * from './NewDisney/NewDisney';
-export * from './Originals/Originals';
-export * from './Trending/Trending';
-export * from './Detail/Detail';
+export * from './Home/HomeElements';
