+++ src/store/details/detail.selectors.ts
@@ -1,3 +1,7 @@
 import {RootState} from '@/store/store';
 
+/**
+ * Селектор для получения деталей фильма по ID
+ */
+
 export const selectMovieDetailsById = (state: RootState) => state.details.data;
