+++ src/store/details/details.actions.ts
@@ -3,6 +3,12 @@ import {doc, getDoc} from 'firebase/firestore';
 import {AppDispatch} from '@/store/store';
 import {movieDetailsActions} from './details.reducer';
 
+/**
+ * Экшн для получения деталей фильма по ID
+ * @param id ID фильма
+ * @returns Promise с параметрами фильма
+ */
+
 export const getMovieDetailsbyId = (id: string) => {
   return (dispatch: AppDispatch) => {
     return getDoc(doc(db, 'movies', id))
