+++ src/store/details/index.ts
@@ -1,0 +1,3 @@
+export * from './details.reducer';
+export * from './details.actions';
+export * from './detail.selectors';
