+++ src/store/movies/index.ts
@@ -1,0 +1,3 @@
+export * from './movies.reducer';
+export * from './movies.actions';
+export * from './movies.selectors';
