+++ src/store/store.ts
@@ -1,5 +1,5 @@
 import {configureStore} from '@reduxjs/toolkit';
-import {rootReducer} from '.';
+import {rootReducer} from '@/store';
 
 const store = configureStore({
   reducer: rootReducer,
